+++ Invoke-Build.ps1
@@ -162,11 +162,11 @@ function Write-Build([ConsoleColor]$Color, [string]$Text) {
 }
 
 #.ExternalHelp Invoke-Build-Help.xml
-function Get-BuildVersion {[Version]'2.6.0'}
+function Get-BuildVersion {[Version]'2.6.1'}
 
 if ($MyInvocation.InvocationName -eq '.') {
 	return @'
-Invoke-Build 2.6.0
+Invoke-Build 2.6.1
 Copyright (c) 2011-2014 Roman Kuzmin
 
 Add-BuildTask (task)